+++ pab/cli.py
@@ -3,7 +3,6 @@ PAB is a framework for developing and running custom tasks in crypto blockchains
 """
 import os
 import sys
-import time
 import inspect
 import getpass
 import logging
@@ -15,7 +14,7 @@ from argparse import ArgumentParser, RawDescriptionHelpFormatter
 
 from pab.blockchain import Blockchain
 from pab.core import PAB
-from pab.config import APP_CONFIG, ENDPOINT, MY_ADDRESS, DATETIME_FORMAT, CONFIG_FILE, DEFAULTS_CONFIG_FILE, KEY_FILE
+from pab.config import APP_CONFIG, ENDPOINT, MY_ADDRESS, DATETIME_FORMAT, CONFIG_FILE, KEY_FILE, override_config_with_defaults, valid_config_found
 from pab.utils import create_keyfile, KeyfileOverrideException
 from pab.alert import alert_exception
 from pab.queue import QueueLoader
@@ -52,11 +51,8 @@ def _create_keyfile(args, logger):
 
 
 def edit_config(args, logger):
-    data = None
-    if not CONFIG_FILE.is_file():
-        with CONFIG_FILE.open("w") as fp:
-            data = DEFAULTS_CONFIG_FILE.open("r").read()
-            fp.write(data)
+    if not valid_config_found():
+        override_config_with_defaults()
     editor = os.environ.get("EDITOR", "vim")
     subprocess.call([editor, CONFIG_FILE])